+++ app/src/test/java/com/zegoggles/smssync/mail/MessageGeneratorTest.java
@@ -19,6 +19,7 @@ import org.mockito.Mock;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
 
+import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
@@ -86,11 +87,11 @@ public class MessageGeneratorTest {
     @Test public void testShouldGenerateSubjectWithNameForMMS() throws Exception {
         PersonRecord personRecord = new PersonRecord(1, "Foo Bar", "foo@bar.com", "1234");
 
-        MmsSupport.MmsDetails details = new MmsSupport.MmsDetails(true, "foo",
-                personRecord,
-                new Address("foo@bar.com"));
+        MmsSupport.MmsDetails details = new MmsSupport.MmsDetails(true, personRecord,
+                Collections.singletonList(personRecord),
+                Collections.singletonList("foo"));
 
-        when(mmsSupport.getDetails(any(Uri.class), any(AddressStyle.class))).thenReturn(details);
+        when(mmsSupport.getDetails(any(Uri.class), any(AddressStyle.class), any(Map.class))).thenReturn(details);
         Message msg = generator.messageForDataType(mockMessage("1234", personRecord), DataType.MMS);
 
         assertThat(msg).isNotNull();
@@ -99,11 +100,11 @@ public class MessageGeneratorTest {
 
     @Test public void testShouldGenerateMMSMessageWithCorrectEncoding() throws Exception {
         PersonRecord personRecord = new PersonRecord(1, "Foo Bar", "foo@bar.com", "1234");
-        MmsSupport.MmsDetails details = new MmsSupport.MmsDetails(true, "foo",
-                personRecord,
-                new Address("foo@bar.com"));
+        MmsSupport.MmsDetails details = new MmsSupport.MmsDetails(true, personRecord,
+                Collections.singletonList(personRecord),
+                Collections.singletonList("foo"));
 
-        when(mmsSupport.getDetails(any(Uri.class), any(AddressStyle.class))).thenReturn(details);
+        when(mmsSupport.getDetails(any(Uri.class), any(AddressStyle.class), any(Map.class))).thenReturn(details);
         Message msg = generator.messageForDataType(mockMessage("1234", personRecord), DataType.MMS);
         assertThat(msg.getHeader(MimeHeader.HEADER_CONTENT_TRANSFER_ENCODING)).isEqualTo(new String[] {
                 MimeUtil.ENC_7BIT
@@ -193,7 +194,7 @@ public class MessageGeneratorTest {
                 any(Map.class),
                 eq(DataType.SMS),
                 anyString(),
-                eq(record),
+                anyString(),
                 eq(date),
                 eq(0));
     }